Adjust the buttons of the header.
