+++ src/components/the-header/the-header.scss
@@ -8,6 +8,12 @@
 
     .HeaderContainer {
         width: 92%;
+
+        .navbar-nav {
+            .nav-item, .btn {
+                margin: 0 0.25rem;
+            }
+        }
     }
 }
