+++ src/views/dashboard/DashboardView.vue
@@ -1,10 +1,12 @@
 <template>
 	<div class='DashboardView'>
-        <h1>Menu</h1>
+        <b-container>
+            <h1>Menu</h1>
         <router-link to='/profile'>My profile</router-link>
 
 		<p>hello</p>
 		<!-- <h1>Hello {{ currentUser.email }}</h1> -->
+        </b-container>
 	</div>
 </template>
